Review README.md
What does "all data" (l.11) refer to? The whole merkle-tree?

What exactly does the subtitle "Component" (l.16) refer to? 

Does "Batcher" subtitle refer to "InMemoryBatcher" (diagram)? Then consistent naming should be used, or define what it refers to in section following subtitle.

"BatchSizeExponent" (l. 20) not consistent with naming in diagram

"EndPoints"? (l. 24) What does that refer to?
 
l. 44-59: Does the GetMerkleProof not need the actual root published to the blockchain to compare with? Or maybe that is what it retrieves using TransactionID?